fix(conceal:limits): group `\limits` with prior command for conceal
commit 11ecdae889d232b3ff4b2acee8eee95b8b578a8c generalized concealing of
`\limits`, but did so at the expense of making it a separate conceal match,
which made it difficult to view/navigate to.
this makes `\limits` concealing match the earlier behavior with `\sum\limits`
and `\lim\limits`, where `limits` is included in the preceding command's conceal
match group, except still preserving generality.
do note that this makes conceal able to display invalid latex expressions where
the preceding command is not a big operator, e.g. `\pi\limits`.