fix: don't see tags with `math` as an equation
In obsidian, the name of a tag is the following hashtag_hashtag-end_meta_tag-<tagname>, if this tag name were to contain the word `math`, then it would recognize the whole tag as an equation instead of a tag. Fix is to check if we are not currently in a tag.
fixes #431.