feat: update docs
safari on iOS should support regex look behind,
source: https://developer.apple.com/documentation/safari-release-notes/safari-16_4-release-notes#JavaScript.
And as such also support it in obsidian. A working version was tested on
ipados 18.

Address #345 where it is not clear what fonts support concealment and
how to add them.

There were multiple threads on forums and a few issues on github saying that they couldn't get snippet files to work.
So I made it extra verbose with examples to make it as clear as possible.
It might be overkill, but better too much than too little.