fix: delayed conceal didn't work on vim
Fixes #429.
codemirror-vim still runs mostly on CM5 code.
Because of that it doesn't deal with decorations correctly. o
Obsidian already fixed the `movingByCharacter` command, but its not clear from the 7k loc where the cursor position logic can be overwritten.

As such to circumvent this bug we update the cursor with `cm.signal("vim-command-done")`.
This causes the vim plugin to "redraw" the cursor.
This does depend on the unsupported `cm.signal` function, which is part of the CM5 API not CM6.
This is a temporary fix until the codemirror-vim plugin is updated to fully support CM6.

This bug only happens with delayedreveal on,
as the decorations update happens after the initial view update and the vim cursor only gets updated directly after the update causing mismatch.