feat: improve snippet performance
`view.dispatch` calls are expensive for large documents,
so a few commits to reduce them.
Improvement 1:
improve snippet performance by removing `view.dispatch` for queueing and
clearing snippets to the snippetQueue. A `view.dispatch` for queueing
and clearing snippets can get up to 10~11ms.
The snippetQueue doesn't need to be in a `StateField` to function,
so now its a regular viewplugin class.

To preserve functionality, the queue is readonly, but no functional
reason I think.