feat: encapsulate snippet variables and snippet triggers in uncaptured group
in light of #447 and #367, it might not be obvious that snippetVariables should be encapsulated in an uncaptured group and since most use cases are \\${snippetVariable}, encapsulting makes writing easier for first time users. Users that somehow need it not encapsulated are probably skilled enough to remove it
and for #472, the snippet trigger should be encapsulated otherwise it might not match at the cursor.