feat: improve snippets performance v3
Improve snippet expansion by reducing `view.dispatch` calls.
Improvement 3:
Merging placing the snippets down and converting the tabstops syntax to actual snippets
To convert the syntax, the changes have to applied separately and
applying the changes to the old tabstops now happens before the new
tabstops are inserted.

The final 2 `view.dispatch` in `expandSnippets` could be merged by
altering the undo and redo behaviour and inserting the keys in the
automatic snippets back in.