fix: detect codebounds better
Fixes #316
Currently when the cursor was at the beginning or at the backticks the cursor would not go over the codeblock's syntaxnodes, instead going over their children and skipping thus the parent.

Also cached the codebounds when the lang is computed as getbounds is going to be used soon after.